refactor: package config auditing with environment
