+++ environments/sv-env-config-verification/sv_env_config_verification/e2_config_auditing/ci/e2_verify_rules.yml
