+++ environments/sv-env-config-verification/sv_env_config_verification/e2_config_auditing/dataset/fixtures/k8s/bad_pod.yaml
