+++ environments/sv-env-config-verification/sv_env_config_verification/e2_config_auditing/dataset/oracle/bad_pod.json
