+++ environments/sv-env-config-verification/sv_env_config_verification/e2_config_auditing/tests/schema_test.py
@@ -1,4 +1,4 @@
-from e2_config_auditing.schema import parse_model_output
+from sv_env_config_verification.e2_config_auditing.schema import parse_model_output
 
 
 def test_parse_model_output_valid() -> None:
